+++ src/runtime/components/DucktoryTabContainer.vue
@@ -5,6 +5,7 @@ const props = withDefaults(defineProps<{
   default: string
   tabClasses?: string
   contentClasses?: string
+  tabs: string[]
 }>(), {
   tabClasses: '',
   contentClasses: '',
@@ -14,10 +15,10 @@ const emit = defineEmits<{
   select: [tab: string]
 }>()
 
-const tab = ref(props.default)
+const active = ref(props.default)
 
 function select(newTab: string) {
-  tab.value = newTab
+  active.value = newTab
   emit('select', newTab)
 }
 </script>
@@ -25,11 +26,18 @@ function select(newTab: string) {
 <template>
   <div class="ducktory-tab-container">
     <div :class="contentClasses">
-      <slot :name="'tab-' + tab" />
+      <template
+        v-for="tab in tabs"
+        :key="tab"
+      >
+        <div v-show="tab === active">
+          <slot :name="'tab-' + tab" />
+        </div>
+      </template>
     </div>
     <div :class="tabClasses">
       <slot
-        :active="tab"
+        :active="active"
         :select="select"
         name="tabs"
       />