+++ src/runtime/pages/story.vue
@@ -1,11 +1,12 @@
 <script lang="ts" setup>
 import { codeToHtml } from 'shiki'
-import { computed, ref, watch } from 'vue'
+import { provide, computed, ref, watch } from 'vue'
 import DucktoryActionBtn from '../components/DucktoryActionBtn.vue'
 import DucktoryTabContainer from '../components/DucktoryTabContainer.vue'
 import { useDucktory } from '../composables/useDucktory'
 import { useHead, useRoute, useRouter } from '#app'
 import type { StoryDefinition } from '~/src/types/StoryDefinition'
+import type { CustomVNode } from '~/src/types/VNodeMagic'
 
 const { stories, getName } = useDucktory()
 const { params, query, path } = useRoute()
@@ -14,12 +15,16 @@ const { push } = useRouter()
 const story = computed(() => stories[params.story as string] ?? null)
 const title = computed(() => story.value ? getName(story.value) : '')
 const pageTitle = computed(() => `Ducktory - ${title.value || 'Not Found'}`)
+const componentDocumentation = ref<CustomVNode[] | undefined>(undefined)
+
+provide('ducktory-documentation', componentDocumentation)
 
 const defaultTab = computed(() => query.tab as string || 'preview')
 
 const codeHighlight = ref('Loading...')
 const justCopied = ref(false)
 
+const hasDocumentation = computed(() => story.value?.meta?.documentation !== undefined || componentDocumentation.value?.length > 0)
 const documentation = computed(() => story.value?.meta?.documentation ?? undefined)
 
 useHead({
@@ -54,13 +59,17 @@ function selectTab(newTab: string) {
 
     <DucktoryTabContainer
       :default="defaultTab"
+      :tabs="['preview', 'code', 'docs']"
       class="ducktory-bg-white ducktory-mt-4 ducktory-overflow-hidden ducktory-shadow-md ducktory-rounded-xl ducktory-p4"
       content-classes="ducktory-p-4"
       tab-classes="ducktory-flex ducktory-border-t ducktory-border-t-gray-200"
       @select="selectTab"
     >
       <template #tab-preview>
-        <component :is="story.componentName" />
+        <component
+          :is="story.componentName"
+          ref="preview"
+        />
       </template>
       <template #tab-code>
         <div
@@ -77,7 +86,19 @@ function selectTab(newTab: string) {
         </div>
       </template>
       <template #tab-docs>
-        <div v-html="documentation" />
+        <ClientOnly>
+          <template v-if="componentDocumentation?.length > 0">
+            <component
+              :is="(node)"
+              v-for="(node, idx) in componentDocumentation"
+              :key="idx"
+            />
+          </template>
+          <div
+            v-else
+            v-html="documentation"
+          />
+        </ClientOnly>
       </template>
 
       <template #tabs="{ select, active }">
@@ -94,7 +115,7 @@ function selectTab(newTab: string) {
           Code
         </DucktoryActionBtn>
         <DucktoryActionBtn
-          v-if="documentation"
+          v-show="hasDocumentation"
           :active="active === 'docs'"
           @click="select('docs')"
         >